Moved benchmark commands to separate file
